clientを別ファイルに切り出し
